# src/main/openapi/spec/20.0.3.yml
openapi: 3.0.2
info:
  title: ${product.name.full} Admin REST API
  description: This is a REST API reference for the ${product.name.full} Admin REST API.
  version: "1"
paths:
...
  /{realm}/clients-initial-access:
    get:
      tags:
        - Client Initial Access
      responses:
        2XX:
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
    post:
      tags:
        - Client Initial Access
      summary: Create a new initial access token.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientInitialAccessCreatePresentation'
        required: true
      responses:
        2XX:
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientInitialAccessPresentation'
    parameters:
      - in: path
        name: realm
        description: realm name (not id!)
        required: true
        schema:
          type: string
        style: simple
  /{realm}/clients-initial-access/{id}:
    delete:
      tags:
        - Client Initial Access
      responses:
        2XX:
          description: success
    parameters:
      - in: path
        name: realm
        description: realm name (not id!)
        required: true
        schema:
          type: string
        style: simple
      - in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
...
components:
  securitySchemes:
    access_token:
      type: http
      scheme: bearer
  schemas:
...
    ClientInitialAccessCreatePresentation:
      type: object
      properties:
        count:
          type: integer
          format: int32
        expiration:
          type: integer
          format: int32
    ClientInitialAccessPresentation:
      type: object
      properties:
        count:
          type: integer
          format: int32
        expiration:
          type: integer
          format: int32
        id:
          type: string
        remainingCount:
          type: integer
          format: int32
        timestamp:
          type: integer
          format: int32
        token:
          type: string
...
security:
  - access_token: []
...